requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
yfinance==0.2.36
plotly==5.18.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
bcrypt==4.1.2
python-jose==3.3.0
requests==2.31.0
ta==0.11.0
ollama==0.1.6 